+++ crates/flipperzero/examples/i2c-ds3231.rs
@@ -14,7 +14,7 @@ extern crate flipperzero_alloc;
 
 use core::ffi::CStr;
 
-use flipperzero::{error, furi::time::Duration, gpio::i2c, println};
+use flipperzero::{error, furi::time::FuriDuration, gpio::i2c, println};
 use flipperzero_rt::{entry, manifest};
 use ufmt::derive::uDebug;
 
@@ -66,7 +66,7 @@ impl RtcTime {
 fn main(_args: Option<&CStr>) -> i32 {
     let mut bus = i2c::Bus::EXTERNAL.acquire();
     let rtc = i2c::DeviceAddress::new(0x68);
-    let timeout = Duration::from_millis(50);
+    let timeout = FuriDuration::from_millis(50);
 
     if bus.is_device_ready(rtc, timeout) {
         let mut data = [0; 7];